Day 1 commit
Here's my EOD Day 1 commits.  I created the bulk of the file structure I
intend to adhere to, and began the process of filling out the
functionalities.  I created the user (seller), merch and review
resources.  As a next step, I need to finish the Index/Show functions on
Merch and nest reviews to sellers.